Fix terrible grid object index system.
I don't know what I was smoking when I designed that thing...

Anyway, now it's more efficient, less complicated, and indices
are guaranteed to be static and not to be reused. Considering
that indices are a primary way for human identification of the
organism everywhere, allowing an organism's index to change
spontaneously is not a great idea.