Modify Default Conflict Handler for reproduction.
The Default Conflict Handler was designed around the critical
assumption that an organism would never generate a conflict that
was not unresolvable with a single move. Adding reproduction,
however, violated this assumption, and so I was forced to modify
the Default Conflict Handler in order to make it versatile enough
that it could handle this new case. I also added a test for this
functionality, and integrated it with the Python code.